Started Build of Flex Navigation
Started to build a navigation based in Flexbox. This dropdown also will have dropdown functionality, and will eventually replace the main navigation entirely. At the moment, just the default (blue) navigation is working, however the functionality is there (minus the dropdown, working on that now). Structure in the example file is also changed, however that will revert when the FN replaces the regular nav.